accounts/views.py
from django.shortcuts import render
from rest_framework.decorators import APIView
from rest_framework.response import Response
from .serializers import RegisterSerializer
from rest_framework import generics
from django.contrib.auth import get_user_model


User = get_user_model()


class RegisterView(generics.CreateAPIView):
    queryset = User.objects.all()
    serializer_class = RegisterSerializer
