requirements.txt
Django==4.2.2
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
asgiref==3.7.2
mysqlclient==2.1.1
Pillow==9.5.0
PyJWT==2.7.0

